// tina-graphql-gateway/apps/demo/.tina/__generated__/types.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** References another document, used as a foreign key */
  Reference: any
  JSON: any
  JSONObject: any
}

export type Node = {
  id: Scalars['ID']
}

export type Document = {
  sys?: Maybe<SystemInfo>
  id: Scalars['ID']
}

export type FormField = {
  label?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  component?: Maybe<Scalars['String']>
}

export type SystemInfo = {
  __typename?: 'SystemInfo'
  filename?: Maybe<Scalars['String']>
  basename?: Maybe<Scalars['String']>
  breadcrumbs?: Maybe<Array<Maybe<Scalars['String']>>>
  path?: Maybe<Scalars['String']>
  relativePath?: Maybe<Scalars['String']>
  extension?: Maybe<Scalars['String']>
  template?: Maybe<Scalars['String']>
  collection?: Maybe<Section>
}
// ...
export type SectionDocumentUnion = Authors_Document | Posts_Document
// ...
export type Authors_Data = Author_Doc_Data

export type Authors_Input = {
  author?: Maybe<Author_Doc_Input>
}

export type Authors_Values = Author_Doc_Values

export type Authors_Form = Author_Doc_Form

export type Authors_Document = Node &
  Document & {
    __typename?: 'Authors_Document'
    id: Scalars['ID']
    sys?: Maybe<SystemInfo>
    data?: Maybe<Authors_Data>
    values?: Maybe<Authors_Values>
    form?: Maybe<Authors_Form>
  }
// ...
export type Posts_Data = Post_Doc_Data

export type Posts_Input = {
  post?: Maybe<Post_Doc_Input>
}

export type Posts_Values = Post_Doc_Values

export type Posts_Form = Post_Doc_Form

export type Posts_Document = Node &
  Document & {
    __typename?: 'Posts_Document'
    id: Scalars['ID']
    sys?: Maybe<SystemInfo>
    data?: Maybe<Posts_Data>
    values?: Maybe<Posts_Values>
    form?: Maybe<Posts_Form>
  }